registroBackend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-jose==3.3.0
cryptography==41.0.0
requests==2.31.0
python-dotenv==1.0.0
